requirements.txt
paho-mqtt
serial